# Force using a buffer from the ChunkRefsPool