**What this PR does / why we need it**:

Re-using a buffer from the `ChunkRefsPool` avoids unnecessary allocations in TSDB's GetChunkRefs call.

**Special notes for your reviewer**:

Benchmarks

```
$ gotest -v -count=1 ./pkg/storage/stores/tsdb/ -test.run="^$" -test.bench=Bench -test.benchtime=5000x
```

```console
$ benchstat tsdbbench-old.txt tsdbbench-new.txt                                                                     
name                      old time/op    new time/op    delta
TenantHeads/10-8             152µs ± 0%      53µs ± 0%   ~     (p=1.000 n=1+1)
TenantHeads/100-8           1.22ms ± 0%    0.39ms ± 0%   ~     (p=1.000 n=1+1)
TenantHeads/1000-8          16.0ms ± 0%     5.2ms ± 0%   ~     (p=1.000 n=1+1)
IndexClient_Stats-8         35.9µs ± 0%    39.3µs ± 0%   ~     (p=1.000 n=1+1)
TSDBIndex_GetChunkRefs-8    2.98ms ± 0%    1.43ms ± 0%   ~     (p=1.000 n=1+1)

name                      old alloc/op   new alloc/op   delta
IndexClient_Stats-8         2.63kB ± 0%    2.14kB ± 0%   ~     (p=1.000 n=1+1)
TSDBIndex_GetChunkRefs-8    6.40MB ± 0%    1.40MB ± 0%   ~     (p=1.000 n=1+1)

name                      old allocs/op  new allocs/op  delta
IndexClient_Stats-8           57.0 ± 0%      57.0 ± 0%   ~     (all equal)
TSDBIndex_GetChunkRefs-8      42.0 ± 0%      27.0 ± 0%   ~     (p=1.000 n=1+1)
```

**Checklist**
- [ ] Reviewed the [`CONTRIBUTING.md`](https://github.com/grafana/loki/blob/main/CONTRIBUTING.md) guide (**required**)
- [ ] Documentation added
- [ ] Tests updated
- [ ] `CHANGELOG.md` updated
- [ ] Changes that require user attention or interaction to upgrade are documented in `docs/sources/upgrading/_index.md`