@bboreham 2023-05-23
Could you run the benchmarks with `-count=5`, or whatever benchstat says you need to get under its significance threshold?