@chaudum 2023-05-24
@bboreham Results with `-count=5`

```console
$ benchstat old.txt new.txt 
name                      old time/op    new time/op    delta
TenantHeads/10-8             127µs ± 6%    110µs ±136%     ~     (p=0.151 n=5+5)
TenantHeads/100-8           1.63ms ±17%    0.58ms ±26%  -64.54%  (p=0.008 n=5+5)
TenantHeads/1000-8          16.3ms ± 4%     6.8ms ± 9%  -58.50%  (p=0.008 n=5+5)
IndexClient_Stats-8         27.0µs ±28%    26.8µs ±30%     ~     (p=0.548 n=5+5)
TSDBIndex_GetChunkRefs-8    3.36ms ± 4%    1.45ms ± 2%  -56.82%  (p=0.008 n=5+5)

name                      old alloc/op   new alloc/op   delta
IndexClient_Stats-8         2.37kB ±12%    2.42kB ±10%     ~     (p=1.000 n=5+5)
TSDBIndex_GetChunkRefs-8    6.40MB ± 0%    1.40MB ± 0%  -78.08%  (p=0.008 n=5+5)

name                      old allocs/op  new allocs/op  delta
IndexClient_Stats-8           59.0 ± 0%      59.0 ± 0%     ~     (all equal)
TSDBIndex_GetChunkRefs-8      42.0 ± 0%      27.0 ± 0%  -35.71%  (p=0.008 n=5+5)
```