@chaudum 2023-05-24
Greater sample (`-count=10`) does not yield much difference:

```console
$ benchstat old.txt new.txt 
name                      old time/op    new time/op    delta
TenantHeads/10-8             244µs ±58%      55µs ± 4%  -77.49%  (p=0.000 n=10+9)
TenantHeads/100-8           1.75ms ± 9%    0.58ms ±26%  -66.89%  (p=0.000 n=9+10)
TenantHeads/1000-8          16.5ms ± 6%     6.4ms ± 2%  -60.89%  (p=0.000 n=9+10)
IndexClient_Stats-8         25.0µs ±10%    25.1µs ±13%     ~     (p=0.863 n=9+9)
TSDBIndex_GetChunkRefs-8    3.34ms ±12%    1.46ms ± 2%  -56.41%  (p=0.000 n=10+9)

name                      old alloc/op   new alloc/op   delta
IndexClient_Stats-8         2.37kB ±12%    2.37kB ±12%     ~     (p=0.923 n=10+10)
TSDBIndex_GetChunkRefs-8    6.40MB ± 0%    1.40MB ± 0%  -78.08%  (p=0.000 n=9+10)

name                      old allocs/op  new allocs/op  delta
IndexClient_Stats-8           59.0 ± 0%      59.0 ± 0%     ~     (all equal)
TSDBIndex_GetChunkRefs-8      42.0 ± 0%      27.0 ± 0%  -35.71%  (p=0.000 n=10+10)
```