@bboreham 2023-05-25
Thanks. I think you might be on an older version of benchstat; it got a bit more chatty recently.