@chaudum 2023-05-25
Output from latest version of `benchstat`

```
goos: linux
goarch: amd64
pkg: github.com/grafana/loki/pkg/storage/stores/tsdb
cpu: 11th Gen Intel(R) Core(TM) i7-1185G7 @ 3.00GHz
                         │    old.txt    │               new.txt                │
                         │    sec/op     │    sec/op     vs base                │
TenantHeads/10-8           244.90µ ± 55%   55.26µ ±  2%  -77.43% (p=0.000 n=10)
TenantHeads/100-8          1735.0µ ± 10%   560.5µ ± 29%  -67.69% (p=0.000 n=10)
TenantHeads/1000-8         16.314m ±  7%   6.414m ±  2%  -60.68% (p=0.000 n=10)
IndexClient_Stats-8         24.89µ ± 10%   24.76µ ± 14%        ~ (p=0.912 n=10)
TSDBIndex_GetChunkRefs-8    3.255m ±  9%   1.453m ±  1%  -55.38% (p=0.000 n=10)
geomean                     891.0µ         372.2µ        -58.22%

                         │    old.txt    │                new.txt                │
                         │     B/op      │     B/op       vs base                │
IndexClient_Stats-8        2.364Ki ± 10%   2.364Ki ± 10%        ~ (p=0.923 n=10)
TSDBIndex_GetChunkRefs-8   6.102Mi ±  0%   1.338Mi ±  0%  -78.08% (p=0.000 n=10)
geomean                    121.5Ki         56.90Ki        -53.18%

                         │  old.txt   │               new.txt                │
                         │ allocs/op  │ allocs/op   vs base                  │
IndexClient_Stats-8        59.00 ± 0%   59.00 ± 0%        ~ (p=1.000 n=10) ¹
TSDBIndex_GetChunkRefs-8   42.00 ± 0%   27.00 ± 0%  -35.71% (p=0.000 n=10)
geomean                    49.78        39.91       -19.82%
¹ all samples are equal
```